requirements.txt
# Основные зависимости для генерации синтетических данных

# Для работы с диффузионными моделями
diffusers>=0.30.0
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0

# Для обработки изображений
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Для ControlNet и depth estimation
# transformers уже включен выше

# Для LoRA (через PEFT, если используется)
peft>=0.8.0

# Для diffusion-pipe (если доступен)
# diffusion-pipe  # Установить отдельно если доступен в виде пакета

# Утилиты
tqdm>=4.66.0

# Опционально: для ускорения (если доступно)
# accelerate>=0.27.0
# xformers>=0.0.23  # Для оптимизации внимания (если доступно)

